# backend/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: MSS Course Platform BFF API
  description: |
    Backend for Frontend API specification for the MSS Course Platform.
    This API defines all endpoints currently implemented using Supabase,
    prepared for migration to Python FastAPI BFF layer.
  version: 1.0.0
  contact:
    name: MSS Course Platform Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-app.vercel.app/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
...
    UserRole:
      type: string
      enum: [user, instructor, admin]
      description: User role in the system
    Profile:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          maxLength: 50
          nullable: true
          description: User's display name (max 50 characters)
        avatar_url:
          type: string
          format: uri
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Category schemas
    Category:
      type: object
      required:
        - id
        - name_ja
        - name_en
        - slug
      properties:
        id:
          type: string
          format: uuid
        name_ja:
          type: string
          description: Japanese name
        name_en:
          type: string
          description: English name
        slug:
          type: string
          description: URL-friendly identifier
        created_at:
          type: string
          format: date-time

    # Course schemas
    CourseBase:
      type: object
      required:
        - title_ja
        - category_id
      properties:
        title_ja:
          type: string
          maxLength: 100
          description: Course title in Japanese
        title_en:
          type: string
          maxLength: 100
          nullable: true
          description: Course title in English
        description_ja:
          type: string
          maxLength: 1000
          nullable: true
          description: Course description in Japanese
        description_en:
          type: string
          maxLength: 1000
          nullable: true
          description: Course description in English
        category_id:
          type: string
          format: uuid
        thumbnail_url:
          type: string
          format: uri
          nullable: true
    Course:
      allOf:
        - $ref: '#/components/schemas/CourseBase'
        - type: object
          required:
            - id
            - instructor_id
            - is_published
          properties:
            id:
              type: string
              format: uuid
            instructor_id:
              type: string
              format: uuid
            is_published:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    CourseDetail:
      allOf:
        - $ref: '#/components/schemas/Course'
        - type: object
          properties:
            categories:
              $ref: '#/components/schemas/Category'
            profiles:
              $ref: '#/components/schemas/Profile'
            videos:
              type: array
              items:
                $ref: '#/components/schemas/Video'
    CourseListItem:
      allOf:
        - type: object
          required:
            - id
            - title_ja
            - created_at
          properties:
            id:
              type: string
              format: uuid
            title_ja:
              type: string
            title_en:
              type: string
              nullable: true
            description_ja:
              type: string
              nullable: true
            description_en:
              type: string
              nullable: true
            thumbnail_url:
              type: string
              format: uri
              nullable: true
            created_at:
              type: string
              format: date-time
            categories:
              type: object
              properties:
                name_ja:
                  type: string
            profiles:
              type: object
              properties:
                email:
                  type: string
            videos:
              type: object
              properties:
                count:
                  type: integer
                  description: Number of videos in the course

    # Video schemas
    VideoBase:
      type: object
      required:
        - youtube_url
        - title_ja
      properties:
        youtube_url:
          type: string
          format: uri
          description: Full YouTube URL
        title_ja:
          type: string
          maxLength: 100
        title_en:
          type: string
          maxLength: 100
          nullable: true
        description_ja:
          type: string
          maxLength: 1000
          nullable: true
        description_en:
          type: string
          maxLength: 1000
          nullable: true
    Video:
      allOf:
        - $ref: '#/components/schemas/VideoBase'
        - type: object
          required:
            - id
            - course_id
            - youtube_video_id
            - order_number
          properties:
            id:
              type: string
              format: uuid
            course_id:
              type: string
              format: uuid
            youtube_video_id:
              type: string
              description: Extracted YouTube video ID
            order_number:
              type: integer
              minimum: 1
              description: Display order within course
            duration:
              type: integer
              nullable: true
              description: Video duration in seconds
            duration_seconds:
              type: integer
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    # Progress schemas
    VideoProgress:
      type: object
      required:
        - id
        - user_id
        - video_id
        - is_completed
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        is_completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProgressSummary:
      type: object
      required:
        - totalVideos
        - completedCount
        - percentage
      properties:
        totalVideos:
          type: integer
        completedCount:
          type: integer
        percentage:
          type: integer
          minimum: 0
          maximum: 100

    # Comment schemas
...
paths:
...
  /courses:
    get:
      tags:
        - Courses
      summary: List published courses
      description: Get all published courses with optional category filter
      operationId: getCourses
      parameters:
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CourseListItem'

    post:
      tags:
        - Courses
      summary: Create a new course
      description: Create a new course (instructor or admin only)
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBase'
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Course'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires instructor or admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{id}:
    get:
      tags:
        - Courses
      summary: Get course details
      description: Get detailed information about a course
      operationId: getCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CourseDetail'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Courses
      summary: Update course
      description: Update course information (owner or admin only)
      operationId: updateCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseBase'
      responses:
        '200':
          description: Course updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Courses
      summary: Delete course
      description: Delete a course (owner or admin only)
      operationId: deleteCourse
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{id}/publish:
    put:
      tags:
        - Courses
      summary: Toggle course publication status
      description: Publish or unpublish a course
      operationId: togglePublish
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - is_published
              properties:
                is_published:
                  type: boolean
      responses:
        '200':
          description: Publication status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /videos/{id}:
    put:
      tags:
        - Videos
      summary: Update video
      description: Update video information
      operationId: updateVideo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoBase'
      responses:
        '200':
          description: Video updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Videos
      summary: Delete video
      description: Delete a video and reorder remaining videos
      operationId: deleteVideo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/me/progress:
    get:
      tags:
        - Progress
      summary: Get user's all progress
      description: Get all video progress for the current user
      operationId: getAllProgress
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoProgress'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/me/courses/{courseId}/progress:
    get:
      tags:
        - Progress
      summary: Get course progress
      description: Get user's progress for a specific course
      operationId: getCourseProgress
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Course progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoProgress'
                  summary:
                    $ref: '#/components/schemas/ProgressSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /videos/{videoId}/progress:
    post:
      tags:
        - Progress
      summary: Update video progress
      description: Mark video as completed or uncompleted
      operationId: updateVideoProgress
      security:
        - bearerAuth: []
      parameters:
        - name: videoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - is_completed
              properties:
                is_completed:
                  type: boolean
      responses:
        '200':
          description: Progress updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/VideoProgress'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Comment endpoints
...